implemented theme chooser  in app settings
