+++ app/src/main/res/menu/main_menu.xml
@@ -3,12 +3,6 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
-        android:id="@+id/action_theme"
-        android:icon="@drawable/ic_theme_24"
-        android:title="@string/select_theme"
-        app:showAsAction="ifRoom" />
-
-    <item
         android:id="@+id/action_logOut"
         android:icon="@drawable/ic_logout_24"
         android:title="@string/logout"
